refactor: extract form model and controller
