+++ test/features/ai/ui/settings/api_key_form_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/features/ai/model/ai_config.dart';
+import 'package:lotti/features/ai/state/api_key_form_controller.dart';
 import 'package:lotti/features/ai/ui/settings/api_key_form.dart';
 
 void main() {
